# Fixing boundary checks on masked grids