--- a/src/kernel/outwards.jl
+++ b/src/kernel/outwards.jl
@@ -34,36 +34,29 @@ Pass grid name `Symbol`s to `R` and `W` type parameters to use specific grids.
 """
 struct OutwardsDispersal{R,W,S<:Stencils.AbstractKernelStencil} <: SetNeighborhoodRule{R,W}
     stencil::S
+    mask::Union{Nothing, Array{Bool}}  # Add an optional mask parameter
 end
+
 function OutwardsDispersal{R,W}(; kw...) where {R,W}
-    OutwardsDispersal{R,W}(DispersalKernel(; kw...))
+    OutwardsDispersal{R,W}(DispersalKernel(; kw...), nothing)
 end
+
+# @inline function applyrule!(data, rule::OutwardsDispersal{R,W}, N, I) where {R,W}
+#     N == zero(N) && return nothing
+#     sum = zero(N)
+#     for (offset, k) in zip(offsets(rule), kernel(rule))
+#         @inbounds propagules = N * k
+#         @inbounds add!(data[W], propagules, I .+ offset...)
+#         sum += propagules
+#     end
+#     @inbounds sub!(data[W], sum, I...)
+#     return nothing
+# end
 
 @inline function applyrule!(data, rule::OutwardsDispersal{R,W}, N, I) where {R,W}
     N == zero(N) && return nothing
-    sum = zero(N)
-    for (offset, k) in zip(offsets(rule), kernel(rule))
-        @inbounds propagules = N * k
-        @inbounds add!(data[W], propagules, I .+ offset...)
-        sum += propagules
-    end
-    @inbounds sub!(data[W], sum, I...)
-    return nothing
-end
-
-# Define the custom OutwardsDispersalWithMask rule
-struct OutwardsDispersalWithMask{R,W,S<:Stencils.AbstractKernelStencil} <: SetNeighborhoodRule{R,W}
-    stencil::S
-end
-
-function OutwardsDispersalWithMask{R,W}(; kw...) where {R,W}
-    OutwardsDispersalWithMask{R,W}(DispersalKernel(; kw...))
-end
-
-@inline function applyrule!(data, rule::OutwardsDispersalWithMask{R,W}, N, I) where {R,W}
-    N == zero(N) && return nothing
     # Check if the current cell is masked, skip if it is
-    if !mask(data)[I...]
+    if rule.mask !== nothing && !rule.mask[I...]
         return nothing
     end
     sum = zero(N)
@@ -71,7 +64,7 @@ end
         target = I .+ offset
         # Check if the target cell is within bounds and not masked
         (target_mod, inbounds) = _inbounds(Reflect(), size(data[1]), target...)
-        if inbounds && mask(data)[target_mod...]
+        if inbounds && (rule.mask === nothing || rule.mask[target_mod...])
             @inbounds propagules = N * k  # Safe to use @inbounds after checks
             @inbounds add!(data[W], propagules, target_mod...)  # Safe to use @inbounds after checks
             sum += propagules