--- a/src/kernel/outwards.jl
+++ b/src/kernel/outwards.jl
@@ -6,7 +6,7 @@
     OutwardsPopulationDispersal{R,W}(; kw...)
 
 Implements deterministic dispersal from the current cell to populations in neighboring
-cells. hello
+cells. ups
 
 This will make sense ecologically where cell populations are large,
 otherwise a randomised kernel may be more suitable.