--- a/src/kernel/outwards.jl
+++ b/src/kernel/outwards.jl
@@ -35,18 +35,19 @@ Pass grid name `Symbol`s to `R` and `W` type parameters to use specific grids.
 struct _Mask end
 const NoMask = nothing
 
-struct OutwardsDispersal{R,W,S<:Stencils.AbstractKernelStencil} <: SetNeighborhoodRule{R,W}
+struct OutwardsDispersal{R,W,S<:Stencils.AbstractKernelStencil, M} <: SetNeighborhoodRule{R,W}
     stencil::S
-    mask_flag::Union{_Mask, Nothing}
+    mask_flag::M
 end
 
-# Constructor for OutwardsDispersal
-function OutwardsDispersal{R,W}(stencil::S; mask_flag::Union{_Mask, Nothing}=NoMask) where {R,W,S<:Stencils.AbstractKernelStencil}
-    OutwardsDispersal{R,W,S}(stencil, mask_flag)
+# Constructors for OutwardsDispersal
+function OutwardsDispersal{R,W}(stencil::S; mask_flag=NoMask) where {R,W,S<:Stencils.AbstractKernelStencil}
+    OutwardsDispersal{R,W,S,typeof(mask_flag)}(stencil, mask_flag)
 end
 
-function OutwardsDispersal{R,W}(; kw...) where {R,W}
-    OutwardsDispersal{R,W}(DispersalKernel(; kw...))
+function OutwardsDispersal{R,W}(; mask_flag=NoMask, kw...) where {R,W}
+    stencil = DispersalKernel(; kw...)
+    OutwardsDispersal{R,W,typeof(stencil),typeof(mask_flag)}(stencil, mask_flag)
 end
 
 @inline function applyrule!(data, rule::OutwardsDispersal{R,W}, N, I) where {R,W}
@@ -71,3 +72,50 @@ end
     @inbounds sub!(data[W], sum, I...)
     return nothing
 end
+
+# @inline function applyrule(data, rule::OutwardsDispersal{R,W}, N, I) where {R,W}
+#     applyrule!(data, rule, N, I)
+# end
+
+################# TESTING #################
+# Define a mask
+mask_data = fill(true, 10, 10)
+for i in 3:9 
+    mask_data[2, i] = false
+    mask_data[3, i] = false
+    mask_data[4, i] = false
+    mask_data[5, i] = false
+    mask_data[6, i] = false
+    mask_data[7, i] = false
+    mask_data[8, i] = false
+end
+
+# Create OutwardsDispersal with and without a mask
+outdisp_with_mask = OutwardsDispersal(
+    formulation=ExponentialKernel(λ=0.0125),
+    distancemethod=AreaToArea(30),
+    mask_flag=_Mask()
+)
+
+outdisp_without_mask = OutwardsDispersal(
+    formulation=ExponentialKernel(λ=0.0125),
+    distancemethod=AreaToArea(30)
+)
+
+# Initialize the simulation grid
+init = fill(100.0, (10,10))
+init[5,5] = 0.0
+
+# Create ruleset and outputs
+rule_with_mask = Ruleset(outdisp_with_mask; boundary=Reflect())
+rule_without_mask = Ruleset(outdisp_without_mask; boundary=Reflect())
+
+# Run the simulation with a mask
+out_with_mask = ArrayOutput(init; tspan=1:100, mask=mask_data)
+a = sim!(out_with_mask, rule_with_mask)
+a[1][5,5]
+a[2][5,5]
+# Run the simulation without a mask
+out_without_mask = ArrayOutput(init; tspan=1:100)
+sim!(out_without_mask, rule_without_mask)
+