--- a/src/kernel/outwards.jl
+++ b/src/kernel/outwards.jl
@@ -29,11 +29,13 @@ is occupied.
     Default is 1.0.
 - `distancemethod`: [`DistanceMethod`](@ref) object for calculating distance between cells.
     The default is [`CentroidToCentroid`](@ref).
+- `mask_flag`: Use `Mask()` to apply masking or `NoMask()` to ignore masking. Default is `NoMask`.
 
 Pass grid name `Symbol`s to `R` and `W` type parameters to use specific grids.
 """
-struct _Mask end
-const NoMask = nothing
+
+struct Mask end
+struct NoMask end
 
 struct OutwardsDispersal{R,W,S<:Stencils.AbstractKernelStencil, M} <: SetNeighborhoodRule{R,W}
     stencil::S
@@ -41,11 +43,11 @@ struct OutwardsDispersal{R,W,S<:Stencils.AbstractKernelStencil, M} <: SetNeighbo
 end
 
 # Constructors for OutwardsDispersal
-function OutwardsDispersal{R,W}(stencil::S; mask_flag=NoMask) where {R,W,S<:Stencils.AbstractKernelStencil}
+function OutwardsDispersal{R,W}(stencil::S; mask_flag::Union{Mask, NoMask}=NoMask()) where {R,W,S<:Stencils.AbstractKernelStencil}
     OutwardsDispersal{R,W,S,typeof(mask_flag)}(stencil, mask_flag)
 end
 
-function OutwardsDispersal{R,W}(; mask_flag=NoMask, kw...) where {R,W}
+function OutwardsDispersal{R,W}(; mask_flag::Union{Mask, NoMask}=NoMask(), kw...) where {R,W}
     stencil = DispersalKernel(; kw...)
     OutwardsDispersal{R,W,typeof(stencil),typeof(mask_flag)}(stencil, mask_flag)
 end
@@ -54,7 +56,7 @@ end
     N == zero(N) && return nothing
     
     # Check if the current cell is masked, skip if it is
-    mask_data = if rule.mask_flag === NoMask nothing else mask(data) end
+    mask_data = if rule.mask_flag === NoMask() nothing else DynamicGrids.mask(data) end
     if !isnothing(mask_data) && !mask_data[I...]
         return nothing
     end
@@ -62,7 +64,7 @@ end
     sum = zero(N)
     for (offset, k) in zip(offsets(rule), kernel(rule))
         target = I .+ offset
-        (target_mod, inbounds) = inbounds(data, target)
+        (target_mod, inbounds) = DynamicGrids.inbounds(data, target)
         if inbounds && (isnothing(mask_data) || mask_data[target_mod...])
             @inbounds propagules = N * k  
             @inbounds add!(data[W], propagules, target_mod...)  
@@ -72,50 +74,3 @@ end
     @inbounds sub!(data[W], sum, I...)
     return nothing
 end
-
-# @inline function applyrule(data, rule::OutwardsDispersal{R,W}, N, I) where {R,W}
-#     applyrule!(data, rule, N, I)
-# end
-
-################# TESTING #################
-# Define a mask
-mask_data = fill(true, 10, 10)
-for i in 3:9 
-    mask_data[2, i] = false
-    mask_data[3, i] = false
-    mask_data[4, i] = false
-    mask_data[5, i] = false
-    mask_data[6, i] = false
-    mask_data[7, i] = false
-    mask_data[8, i] = false
-end
-
-# Create OutwardsDispersal with and without a mask
-outdisp_with_mask = OutwardsDispersal(
-    formulation=ExponentialKernel(λ=0.0125),
-    distancemethod=AreaToArea(30),
-    mask_flag=_Mask()
-)
-
-outdisp_without_mask = OutwardsDispersal(
-    formulation=ExponentialKernel(λ=0.0125),
-    distancemethod=AreaToArea(30)
-)
-
-# Initialize the simulation grid
-init = fill(100.0, (10,10))
-init[5,5] = 0.0
-
-# Create ruleset and outputs
-rule_with_mask = Ruleset(outdisp_with_mask; boundary=Reflect())
-rule_without_mask = Ruleset(outdisp_without_mask; boundary=Reflect())
-
-# Run the simulation with a mask
-out_with_mask = ArrayOutput(init; tspan=1:100, mask=mask_data)
-a = sim!(out_with_mask, rule_with_mask)
-a[1][5,5]
-a[2][5,5]
-# Run the simulation without a mask
-out_without_mask = ArrayOutput(init; tspan=1:100)
-sim!(out_without_mask, rule_without_mask)
-