--- a/src/kernel/outwards.jl
+++ b/src/kernel/outwards.jl
@@ -56,7 +56,7 @@ end
     N == zero(N) && return nothing
     
     # Check if the current cell is masked, skip if it is
-    mask_data = if rule.mask_flag === NoMask() nothing else DynamicGrids.mask(data) end
+    mask_data = rule.mask_flag === NoMask() ? nothing : DynamicGrids.mask(data)
     if !isnothing(mask_data) && !mask_data[I...]
         return nothing
     end