@rafaqz 2024-06-13
Ok instead of attaching the mask to the rule, we can just get it from `data` with `mask(data)`.

We can have the rule field as `nothing` and some other value? maybe a singleton `_Mask()`? Then in the constructor for the rule we switch `true/false` to `_Mask()`/`nothing`.

This is better than the field being `true` or `false`, because the compiler cant remove the check for true/false, but it can for two different types.