@rafaqz 2024-06-13
`kernelproduct` is in Stencils.jl now (try `@which kernelproduct`).

Thats available to Dispersal.jl, so it just works. 

(we need `kernelproduct` because `Base.dot` is recursive, which is awful)