@nicolasmerino41 2024-06-13
But how will mask(data) work when it's not missingval's what's in the masked area? And I don't think using the `with=` keyword would work either cause it'd need to be specified before.