@nicolasmerino41 2024-06-13
Well, I'll leave it for today. I don't even know what I'm doing at this point jajaja. I created what I think it's a singleton `struct _Mask end`  so it can be specified when using OutwardsDispersal, otherwise mask_flag is nothing by default. 

Then I added the mask(data) approach within `applyrule!` but still don't understand why it should work, which it doesn't so far :)

```
@inline function applyrule!(data, rule::OutwardsDispersal{R,W}, N, I) where {R,W}
    N == zero(N) && return nothing
    
    # Check if the current cell is masked, skip if it is
    mask_data = if rule.mask_flag === NoMask nothing else mask(data) end
    if !isnothing(mask_data) && !mask_data[I...]
        return nothing
    end
    
    sum = zero(N)
    for (offset, k) in zip(offsets(rule), kernel(rule))
        target = I .+ offset
        (target_mod, inbounds) = inbounds(data, target)
        if inbounds && (isnothing(mask_data) || mask_data[target_mod...])
            @inbounds propagules = N * k  
            @inbounds add!(data[W], propagules, target_mod...)  
            sum += propagules
        end
    end
    @inbounds sub!(data[W], sum, I...)
    return nothing
end
```