@rafaqz 2024-06-13
I mean `DynamicGrids.mask`! That's what you will use inside DynamicGrids.jl, as theres no Rasters.jl dep.

Its unfortunate the name is the same, its from before Rasters existed. It just gets the mask object that you pass in to the output.