@nicolasmerino41 2024-06-13
That makes much more sense!

On Thu, 13 Jun 2024, 19:11 Rafael Schouten, ***@***.***>
wrote:

> I mean DynamicGrids.mask! That's what you will use inside
> DynamicGrids.jl, as theres no Rasters.jl dep.
>
> Its unfortunate the name is the same, its from before Rasters existed. It
> just gets the mask object that you pass in to the output.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/cesaraustralia/Dispersal.jl/pull/114#issuecomment-2166343318>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ALXYX3YAAYR6JS5EDNVTXI3ZHHHE3AVCNFSM6AAAAABJIEWBZWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDCNRWGM2DGMZRHA>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>