@nicolasmerino41 2024-06-14
Good morning. It's all good now. It works perfectly. I took the freedom to write a test for it but you might find it trivial so feel free to tell me to erase it. 

It works perfect within the package but I can't use it in another environment. I guess I'm missing some definition? Any idea on that?

Also, I included Stencils in deps, I don't know if that's correct but I was struggling to make use of it without being there, let me know :)