@rafaqz 2024-06-14
Amazing!

And trivial tests are actually great. The main idea is to clearly demonstrate that obvious things do work exactly as we intend. Especially that your problem is solved - the populations are conserved over many iterations in the simulation.

Its fine to have Stencils in the deps, to be more explicit. But you can also do `using DynamicGrids.Stencils`.

Looks like I will have to work on the tests here too to get them passing, I only did DynamicGrids.jl so far.


The last thing for you now (if you havent already) is to make sure things are still fast, and not too much slower with the mask check. You might want to look at a ProfileView.jl profile as well, to make sure there are no type instability coming from the rule. Paste timings and /or profile images here if you do any.