@nicolasmerino41 2024-06-14
True! I forgot to benchmark it. I'll get to it. 
I was confused with DynamicGrids.Stencils since I thought that was referring to when stencils was part of DG and since I was using the new DG#dev, I assumed stencils feature would come from Stencils.jl but then it didn't work without also including DynamicGrids.Stencils, so quite confusing to me jeje