@nicolasmerino41 2024-06-15
Hey there,

I just ran the benchmarks, sorry for the delay :)
So for the default package (before the changes):
```
# Define a mask
mask_data = [i == 1 || i == 100 || j == 1 || j == 100 ? false : true for i in 1:100, j in 1:100]

# Create a grid with empty borders matching the mask
init = map(x -> x ? 100.0 : 0.0, mask_data)

outdisp = OutwardsDispersal(
    formulation=Dispersal.ExponentialKernel(λ=0.0125),
    distancemethod=AreaToArea(30)
)
# Run the simulation with a mask
output_with_mask = ArrayOutput(init; tspan=1:1000, mask=mask_data)
@btime sim!(output_with_mask, Ruleset(outdisp; boundary=Reflect()))

# Run the simulation without a mask
output_without_mask = ArrayOutput(init; tspan=1:1000)
@btime sim!(output_without_mask, rule_without_mask)
```
This gave 51ms for the masked simulation and 43ms for the one without.

Now, I ran the exact same code with the new PR implementation, the same two given mask_flag = NoMask is default and an additional one with Mask(), i.e:

```
# Define a mask
mask_data = [i == 1 || i == 100 || j == 1 || j == 100 ? false : true for i in 1:100, j in 1:100]

# Create a grid with empty borders matching the mask
init = map(x -> x ? 100.0 : 0.0, mask_data)

# Create OutwardsDispersal without a mask, NoMask is default
outdisp_without_mask = OutwardsDispersal(
    formulation=ExponentialKernel(λ=0.0125),
    distancemethod=AreaToArea(30)
)

# Run the simulation with a mask
output_with_mask = ArrayOutput(init; tspan=1:1000, mask=mask_data)
@btime sim!(output_with_mask, Ruleset(outdisp_without_mask; boundary=Reflect()))

# Run the simulation without a mask
output_without_mask = ArrayOutput(init; tspan=1:1000)
@btime sim!(output_without_mask, Ruleset(outdisp_without_mask; boundary=Reflect()))

## TESTING NEW IMPLEMENTATION
# Create OutwardsDispersal with a mask
outdisp_with_mask = OutwardsDispersal(
    formulation=Dispersal.ExponentialKernel(λ=0.0125),
    distancemethod=AreaToArea(30),
    mask_flag=Dispersal.Mask()
)

# Run the simulation with a mask
output_with_mask_and_masked_rule = ArrayOutput(init; tspan=1:1000, mask=mask_data)
@btime sim!(output_with_mask_and_masked_rule, Ruleset(outdisp_without_mask; boundary=Reflect()))
```

And I got 59ms, 54ms and 127s, respectively. So it looks like using the new feature does slow down the process a lil bit. Fortunately, the implementation only caused a tiny slowing when you choose NOT to use the new feature. We can notify the perfomance drop caused by Mask() so the user is aware. Here I attach a ProfileView screenshot, I've tried to interprete it but it links to many scripts from who knows where, perhaps the amount of red lines gives you a hint. 
![ProfileView screenshot](https://github.com/cesaraustralia/Dispersal.jl/assets/49253359/2d0c9bea-6060-4bbe-ae40-97c1fcb08776)

Have good weekend :)
Nico

Ps: For perspective, the first two trials using InwardsDispersal are 33ms and 15ms, for masked and unmasked grid, respectively.