@rafaqz 2024-06-15
Ah interesting. Yes there should be no cost of not using the mask.

That profile is likely the first run, it looks like compilation? Maybe run it a few times ;)