@nicolasmerino41 2024-06-17
Ok, for perspective here's a full benchmarking for the original package (O), my script (M) and your suggestions(S):
With mask but OutwardsDispersal NoMask(): O:75ms, M:190ms, S:197ms
Without mask and OutwardsDispersal NoMask(): O:66ms, M:186, S:193ms
With mask and OutwardsDispersal Mask(): O:NoFeature, M:260ms, S:260ms
Without mask and InwardsDispersal: 20ms
With mask and InwardsDispersal: 50ms
I guess if you'd like to go back to original performance we'd have to take the route of creating a completely different function, like OutwardsDispersalWithMask? so you can keep the performance of the original OutwardsDispersal through not checking any mask at all